Sound ids are SfxName, not i32
MobjInfo's five sound fields and s_start_sound take SfxName, so the 860
'SfxName::X as i32' casts (mostly the info.rs tables) are gone. The four
'sfx_posit1 + P_Random() % 3' sites pick from an array of the variants
(one P_Random call each, same as before); '!= 0' sentinels are SfxNone.

+++ rust-doomgeneric/engine/src/g_game.rs
@@ -1133,7 +1133,7 @@ pub fn check_spot(state: &mut GameState, playernum: i32, mthing: &MapThing) -> b
         .floorheight;
     let mo = spawn_mobj(state, x + 20 * xa, y + 20 * ya, floorheight, MobjType::Tfog);
     if state.game.g_game.players[state.game.g_game.consoleplayer].viewz != 1 {
-        s_start_sound(state, SoundOrigin::Mobj(mo), SfxName::Telept as i32);
+        s_start_sound(state, SoundOrigin::Mobj(mo), SfxName::Telept);
     }
     true
 }
